# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Plataforma de Reportes Comunitarios - ResQNET
  description: API para gestionar reportes de seguridad y emergencias en la comunidad.
  version: 1.0.0
servers:
  - url: http://localhost:9090/
    description: Servidor local
paths:
...
  /users:
    post:
      summary: User registration
      description: Allows users to register by providing their personal data.
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationDTO'
      responses:
        '201':
          description: User registered successfully.
        '400':
          description: Invalid data.
        '409':
          description: The provided email is already registered.
        '500':
          description: Internal server error.
        '503':
          description: Service unavailable.

    get:
      summary: Get paginated list of users
      tags:
        - Users
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
          description: The page number to retrieve (starting from 0).
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
          description: The number of users per page (max 100).
      responses:
        "200":
          description: Paginated user list retrieved successfully.
        "401":
          description: Unauthorized - User is not authenticated.
        "403":
          description: Forbidden - User does not have permission to access this resource.
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.
  /users/{id}:
    get:
      summary: Get user information
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user to retrieve.
      responses:
        '200':
          description: User information retrieved successfully.
        '400':
          description: Invalid user ID.
        '404':
          description: User not found.
        '500':
          description: Internal server error.

    put:
      summary: Update user
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationDTO'
      responses:
        '204':
          description: User updated successfully.
        '400':
          description: Invalid request data.
        '404':
          description: User not found.
        '500':
          description: Internal server error.

    delete:
      summary: Delete user
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user to delete.
      responses:
        '204':
          description: User deleted successfully.
        '400':
          description: Invalid user ID.
        '404':
          description: User not found.
        '500':
          description: Internal server error.

  # Authentication
...
components:
  schemas:
    UserRegistrationDTO:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          pattern: '^\+?[0-9]{10,15}$'
        city:
          type: string
          minLength: 2
          maxLength: 100
        address:
          type: string
          minLength: 5
          maxLength: 200
        password:
          type: string
          minLength: 8
          maxLength: 50
...